scoring

+++ Birds and Bees/Assets/Rufino/Scripts/GameManager.cs
@@ -7,7 +7,7 @@ public class GameManager : MonoBehaviour {
 
 	//0 is NO player selected
 	public static int[] players = new int[] {1,2,5,4};
-	public static int[] score = new int[] {0,0,20,0};
+	public static float[] score = new float[] {0,0,20,0};
 	
 	// Use this for initialization
 	void Start () {
